refactor(bpane-client): split public bpane facade
